main.js разбит на модули, подключен к index.html
